requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.1.0
arrow==1.2.3
asciitree==0.3.3
asn1crypto==0.24.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blessed==1.20.0
boto==2.49.0
boto3==1.26.88
botocore==1.29.88
bravado==11.0.3
bravado-core==5.17.1
Brlapi==0.6.6
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.1.0
comm==0.1.2
configmypy==0.1.0
crcmod==1.7
croniter==1.3.8
cryptography==37.0.2
cupshelpers==1.0
cycler==0.10.0
dateutils==0.6.12
debugpy==1.4.1
decorator==5.0.9
deepdiff==6.3.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
distro-info==0.18ubuntu0.18.04.1
dnspython==2.3.0
docker-pycreds==0.4.0
email-validator==1.3.1
entrypoints==0.3
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.88.0
fasteners==0.17.3
filelock==3.4.2
frozenlist==1.3.0
fsspec==2023.3.0
future==0.18.3
gcs-oauth2-boto-plugin==3.0
gdown==4.2.1
gitdb==4.0.10
GitPython==3.1.31
google-apitools==0.5.32
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-reauth==0.1.1
grpcio==1.44.0
gsutil==5.10
gym==0.24.1
gym-notices==0.0.7
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httplib2==0.20.4
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.1
idna==3.2
importlib-metadata==4.11.2
iniconfig==2.0.0
inquirer==3.1.3
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.0.1
json5==0.9.6
jsonpointer==2.3
jsonref==1.1.0
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
jupyterlab-widgets==1.0.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.2
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
learn2learn==0.1.7
lightning==2.0.1
lightning-cloud==0.5.32
lightning-utilities==0.7.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mdurl==0.1.2
mistune==0.8.4
monotonic==1.6
msgpack==1.0.5
multidict==6.0.2
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
neptune-client==1.0.2
nest-asyncio==1.5.1
netifaces==0.10.4
nose==1.3.7
notebook==6.4.3
numcodecs==0.11.0
numpy==1.21.2
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.45.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.9
packaging==21.0
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==23.0.1
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.26.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.2.2
pytest-mock==3.10.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2021.1
pyu2f==0.1.5
pyxdg==0.25
PyYAML==6.0
pyzmq==25.0.0
qpth==0.0.15
readchar==4.0.5
reportlab==3.4.0
requests==2.26.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retry-decorator==1.1.1
rfc3986==1.5.0
rfc3987==1.3.8
rich==13.3.3
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
sentry-sdk==1.17.0
setproctitle==1.3.2
setuptools==58.2.0
simplejson==3.13.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
ssh-import-id==5.7
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
strict-rfc3339==0.7
swagger-spec-validator==3.0.3
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorly==0.8.1
tensorly-torch==0.3.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==3.0.0
timm==0.6.12
tomli==2.0.1
torch==1.11.0+cu113
torch-summary==1.4.5
torchaudio==0.11.0+cu113
torchmetrics==0.11.3
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.8.1
typing_extensions==4.5.0
ufw==0.36
ujson==5.7.0
urllib3==1.26.15
usb-creator==0.3.3
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.13.2
wadllib==1.3.2
wandb==0.14.0
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.2.1
websockets==11.0.1
Werkzeug==2.0.3
wheel==0.30.0
widgetsnbextension==3.5.2
xkit==0.0.0
yarl==1.7.2
zarr==2.14.2
zipp==3.7.0
zope.interface==4.3.2
nose2==0.14.0
